deleted unnecessary commented code and removed redundant imports
