+++ exconvnet/des_data/__init__.py
@@ -1,2 +1,1 @@
-# init
 from .fetcher import fetch
